New bot advisor :) working well
